requirements.txt
Pillow==4.1.0
pyTelegramBotAPI==2.3.1
